🎨 Change alert to modal
