+++ src/pages/reset-password/change-password/index.tsx
@@ -1,7 +1,12 @@
+import SignLayout from '@/src/components/common/layout/Signlayout';
 import ChangePwContent from '@/src/components/resetPwContent/changePwContent';
 
 const ChangePassword = () => {
-  return <ChangePwContent />;
+  return (
+    <SignLayout>
+      <ChangePwContent />
+    </SignLayout>
+  );
 };
 
 export default ChangePassword;
