+++ src/pages/reset-password/index.tsx
@@ -1,7 +1,12 @@
+import SignLayout from '@/src/components/common/layout/Signlayout';
 import ResetPwContent from '@/src/components/resetPwContent';
 
 const ResetPw = () => {
-  return <ResetPwContent />;
+  return (
+    <SignLayout>
+      <ResetPwContent />
+    </SignLayout>
+  );
 };
 
 export default ResetPw;
